Assume edition 0. Also R1 for push to freenet://
Edition is easier to omit, and having Fred search for the latest edition
is a reasonable default. When using internal commands (fn-*) the user
is assumed to be more confident in Infocalypse internals.

When using built-in commands with freenet: URIs, redundancy is less
clear, so assuming redundancy seems helpful in use cases like
hg clone . freenet:USK@.../repo_path. This also adds a warning when
creating a repository on a path that does not specify redundancy.

Assume edition:
* fn-create --uri and --wot
* fn-copy --inserturi and --requesturi
* fn-reinsert --uri
* fn-pull --uri
* fn-push --uri
* built-in command source

Assume edition and redundancy:
* built-in command destination