Track the latest known repo list edition.
This takes some responsibility off the node for fetching an up-to-date
edition, because it is no longer always given edition 0 to start from.
This does introduce some clumsiness as if the node follows redirects
it does not return the used edition. This should probably be replaced
with a WoT property once WoT properties can be inserted without causing
an insert, as it will avoid starting the repo list at edition 0.